細かな気になってた奴直した

+++ Manasona4/source/UI/SPGage.cpp
@@ -188,7 +188,7 @@ void SpGage::Render()
 				m_iGageWidth - l_w, m_iGageHeight*m_iGageLevel,
 				l_w, m_iGageHeight, RS::ADD);
 
-			int number = m_pPlayerReferences->GetOverDriveGage();
+			int number = (int)m_pPlayerReferences->GetOverDriveGage();
 			for (int count = 0; count <= 2; ++count)
 			{
 				int digitNum = number % 10;	// ˆê”Ô¬‚³‚¢Œ…‚ð“üŽè
@@ -246,7 +246,7 @@ void SpGage::Render()
 				0, m_iGageHeight*m_iGageLevel,
 				l_w, m_iGageHeight, RS::ADD);
 
-			int number = m_pPlayerReferences->GetOverDriveGage();
+			int number = (int)m_pPlayerReferences->GetOverDriveGage();
 			for (int count = 0; count <= 2; ++count)
 			{
 				int digitNum = number % 10;	// ˆê”Ô¬‚³‚¢Œ…‚ð“üŽè
